refactor(netx): make resolver public
This will allow me to cherry-pick specific bits of the resolver once
the code has been further refactored to use the decorator pattern.

Part of https://github.com/ooni/probe-engine/issues/359
