+++ netx/dialer.go
@@ -12,8 +12,8 @@ import (
 
 	"github.com/ooni/probe-engine/netx/dialer"
 	"github.com/ooni/probe-engine/netx/handlers"
-	"github.com/ooni/probe-engine/netx/internal/resolver"
 	"github.com/ooni/probe-engine/netx/modelx"
+	"github.com/ooni/probe-engine/netx/resolver"
 )
 
 // Dialer performs measurements while dialing.
